Added ability to insert default exercise. Updated BaseDao and few repositories methods.

+++ data/src/main/kotlin/com/dvinc/workouttimer/data/database/entity/exercise/ExerciseEntity.kt
@@ -21,7 +21,7 @@ data class ExerciseEntity(
         var uid: Int = 0,
 
         @ColumnInfo(name = "workout_id")
-        val workoutId: Int,
+        val workoutId: Long,
 
         @ColumnInfo(name = "name")
         val name: String,
