+++ domain/src/main/kotlin/com/dvinc/workouttimer/domain/model/exercise/Exercise.kt
@@ -7,7 +7,7 @@ package com.dvinc.workouttimer.domain.model.exercise
 
 data class Exercise(
         val id: Int,
-        val workoutId: Int,
+        val workoutId: Long,
         val name: String,
         val description: String,
         val time: Long,
